Make Detecting Test Failures Easier
Adds code to PTdFgenTester.py that makes detecting and diagnosing failed tests
easier. Adds an announceTest() function that prints what test is running.
Rearranges the comments at the top of each test so that the order corresponds
to the order the files are tested, making it easier to match up the error
messages printed.

Changes:

 * tests/PTdFgenTester.py: Add announceTest() function and rearrange comments.